Dodat alert za brisanje kursa

+++ src/app/course/course.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import Swal from 'sweetalert2';
 import { Course } from '../model/course';
 import { CourseService } from '../service/course.service';
 
@@ -29,9 +30,33 @@ export class CourseComponent implements OnInit {
   }
 
   deleteCourse(id: number) {
-    this.courseService.deleteCourse(id).subscribe(data=> {
-      console.log(data);
-      this.getCourses();
-    },error => console.error(error));
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.courseService.deleteCourse(id).subscribe(data=> {
+            console.log(data);
+          Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+          );
+          this.getCourses();
+          },error => {
+            console.error(error);
+            Swal.fire({
+              icon: 'error',
+              title: 'Oops...',
+              text: 'Something went wrong!'
+            });
+      });
+      }
+    })
   }
 }
